Consolidated the auto install tests. Fixed up TestedProductRule definition in AbstractFileBasedServerTest.java so types were working corr ectly and visit could be called on TestedProductRule.

+++ atlassian-webdriver/atlassian-webdriver-core/src/main/java/com/atlassian/webdriver/testing/rule/SessionCleanupRule.java
@@ -13,6 +13,9 @@ public class SessionCleanupRule extends ExternalResource
     protected void after()
     {
         AtlassianWebDriver driver = LifecycleAwareWebDriverGrid.getCurrentDriver();
-        driver.manage().deleteAllCookies();
+        if (driver != null)
+        {
+            driver.manage().deleteAllCookies();
+        }
     }
 }
